refactor : props interface 정리
